[UI] - Adding characters image to list and details - Adapter Refactor
