+++ .idea/misc.xml
@@ -5,7 +5,7 @@
       <map>
         <entry key="app/src/main/res/layout/activity_detail.xml" value="0.2171875" />
         <entry key="app/src/main/res/layout/activity_list.xml" value="0.2171875" />
-        <entry key="app/src/main/res/layout/item_character.xml" value="0.2171875" />
+        <entry key="app/src/main/res/layout/item_character.xml" value="0.67" />
       </map>
     </option>
   </component>
