+++ app/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
+    id 'kotlin-kapt'
 }
 
 
@@ -40,10 +41,15 @@ android {
 dependencies {
     implementation "androidx.appcompat:appcompat:$app_compat_version"
     implementation "androidx.recyclerview:recyclerview:$recycler_view_version"
+    implementation"androidx.constraintlayout:constraintlayout:$constraint_layout_version"
 
     implementation project(':domain')
     implementation project(':data') //only used for networkModule
 
     // Koin Core features
     implementation "io.insert-koin:koin-android:$koin_version"
+
+    //Glide
+    implementation "com.github.bumptech.glide:glide:$glide_version"
+    annotationProcessor "com.github.bumptech.glide:compiler:$glide_version"
 }
