+++ buildsystem/dependencies.gradle
@@ -16,6 +16,8 @@ ext {
     //Presentation
     app_compat_version = "1.4.2"
     recycler_view_version = "1.2.1"
+    glide_version = "4.13.0"
+    constraint_layout_version = "2.1.4"
 
     //test
     junit_version = "4.13.2"
